+++ src/IResultsFormatter.php
@@ -3,37 +3,15 @@ declare(strict_types=1);
 
 namespace MyTester;
 
+use Konecnyjakub\EventDispatcher\IEventSubscriber;
+
 /**
  * Results formatter for {@see Tester}
  *
  * @author Jakub Konečný
  */
-interface IResultsFormatter
+interface IResultsFormatter extends IEventSubscriber
 {
-    /**
-     * Report that we started running tests
-     *
-     * @param TestCase[] $testCases
-     */
-    public function reportTestsStarted(array $testCases): void;
-
-    /**
-     * Report that all tests finished
-     *
-     * @param TestCase[] $testCases
-     */
-    public function reportTestsFinished(array $testCases): void;
-
-    /**
-     * Report that a {@see TestCase} was started
-     */
-    public function reportTestCaseStarted(TestCase $testCase): void;
-
-    /**
-     * Report results of one {@see TestCase}
-     */
-    public function reportTestCaseFinished(TestCase $testCase): void;
-
     /**
      * Generates results of Tester run and outputs it to set file/console
      *