+++ src/ResultsFormatters/Console.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 namespace MyTester\ResultsFormatters;
 
 use MyTester\ConsoleColors;
+use MyTester\Events\TestCaseFinished;
 use MyTester\IConsoleAwareResultsFormatter;
 use MyTester\JobResult;
 use MyTester\SkippedTest;
-use MyTester\TestCase;
 use MyTester\TestWarning;
 
 /**
@@ -36,10 +36,9 @@ final class Console extends AbstractResultsFormatter implements IConsoleAwareRes
         $this->console = $console;
     }
 
-    public function reportTestCaseFinished(TestCase $testCase): void
+    public function reportTestCaseFinished(TestCaseFinished $event): void
     {
-        $jobs = $testCase->jobs;
-        foreach ($jobs as $job) {
+        foreach ($event->testCase->jobs as $job) {
             switch ($job->result) {
                 case JobResult::SKIPPED:
                     $this->skipped[] = new SkippedTest($job->name, (is_string($job->skip) ? $job->skip : ""));